factor out LastSync component
