+++ client/src/pages/dashboard/sidebar.tsx
@@ -1,12 +1,14 @@
-import React, { useState } from "react"
+import React, { useEffect, useState } from "react"
 import { Link as RouterLink } from "react-router-dom"
 import { toast } from "react-hot-toast"
 
+import { formatTimeAgo } from "../../util/misc"
+import Spinner from "../../components/spinner"
 import api from "../../util/api"
 import ConversationsList from "./conversations_list"
 import { BiSolidBarChartAlt2 } from "react-icons/bi"
 import { ListingSelector } from "./listing_selector"
-import { Box, Flex, Text } from "theme-ui"
+import { Box, Flex, Text, Link } from "theme-ui"
 import { TbFocus } from "react-icons/tb"
 
 const LogoBlock = () => {
@@ -32,9 +34,41 @@ const LogoBlock = () => {
   )
 }
 
+const LastSync = ({ lastSync, syncInProgress, syncPRs }: { lastSync: number; syncInProgress: boolean; syncPRs: () => void }) => {
+  return <Box
+    sx={{
+      position: "absolute",
+      bottom: "0px",
+      pb: "86px",
+      pt: "20px",
+      fontSize: "0.88em",
+      textAlign: "center",
+      width: "100%",
+      color: "#83817d",
+      background:
+        "linear-gradient(0deg, #faf9f6 0%, #faf9f6 66%, #faf9f699 88%, transparent)",
+    }}
+  >
+    Last sync: {isNaN(lastSync) ? "n/a" : formatTimeAgo(lastSync)} &nbsp;
+    <Link variant="links.a" onClick={() => syncPRs()}>
+      {syncInProgress ? <Spinner size={26} /> : `Sync now`}
+    </Link>
+  </Box>
+}
+
 
 const Sidebar = ({mobileMenuOpen}: {mobileMenuOpen: boolean}) => {
   const [syncInProgress, setSyncInProgress] = useState(false)
+  const [lastSync, setLastSync] = useState<number>()
+
+  useEffect(() => {
+    api.get("/api/v3/github_syncs", {}).then((result) => {
+      if (!result.success) return
+      const lastSync = Date.parse(result.latest.ts)
+      setLastSync(lastSync)
+    })
+  }, [])
+
   const syncPRs = () => {
     // github sync
     setSyncInProgress(true)
@@ -97,6 +131,7 @@ const Sidebar = ({mobileMenuOpen}: {mobileMenuOpen: boolean}) => {
       syncPRs={syncPRs}
       syncInProgress={syncInProgress}
     />
+    <LastSync lastSync={lastSync} syncInProgress={syncInProgress} syncPRs={syncPRs} />
   </Box>
 }
